update file script

+++ reach_exp_lsb.py
@@ -1,20 +1,15 @@
 import os.path
-import signal
+import subprocess
 import sys
 
-from mono_proof import run_and_prove, reset, Record
+from mono_proof import  Record
 from glob import glob
 
 from parser import reextension
 
 
-def signal_handler(signum, frame):
-    if signum == signal.SIGALRM:
-        print("timeout {}".format(frame))
-        raise TimeoutError
-instance_timeout = 5000
 
-signal.signal(signal.SIGALRM, signal_handler)
+instance_timeout = 3600
 
 
 if __name__ == "__main__":
@@ -45,30 +40,46 @@ if __name__ == "__main__":
 
     test_files = glob("{}/**/*.gnf".format(input_directory), recursive=True)
     with open(output_csv, 'w') as o_file:
-        r = Record("test")
-        o_file.write("{}\n".format(r.print_header()))
+        # r = Record("test")
+        # o_file.write("{}\n".format(r.print_header()))
         for file in test_files:
             print(file)
-            r = Record(os.path.basename(file))
-            signal.alarm(instance_timeout)
+            arugment_list = ["python3", "run_mono_proof.py", file, "--no-backward-check", "--lemma-bitblast", "--no-graph-reduction",
+                             "--no-witness-reduction"]
+
+            process = subprocess.Popen(arugment_list,
+                                       stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
+
             try:
-                run_and_prove(file, r, running_opt=[], witness_reduction=True, backward_check=backward_check,
-                              lemma_bitblast=lemma_bitblast, graph_reduction=graph_reduction)
-                o_file.write(str(r) + '\n')
-            except TimeoutError:
-                o_file.write("{} timeout ({} secs) \n".format(file, instance_timeout))
-            except Exception as e:
-                print(e)
-                o_file.write("{} error) \n".format(file, instance_timeout))
+                stdout, stderr = process.communicate(timeout=instance_timeout)
+            except subprocess.TimeoutExpired:
+                process.kill()
+            #
+            # print(file)
+            # r = Record(os.path.basename(file))
+            # signal.alarm(instance_timeout)
+            # try:
+            #     run_and_prove(file, r, running_opt=[], witness_reduction=True, backward_check=backward_check,
+            #                   lemma_bitblast=lemma_bitblast, graph_reduction=graph_reduction)
+            #     o_file.write(str(r) + '\n')
+            # except TimeoutError:
+            #     o_file.write("{} timeout ({} secs) \n".format(file, instance_timeout))
+            # except Exception as e:
+            #     print(e)
+            #     o_file.write("{} error) \n".format(file, instance_timeout))
             finally:
-                # reset alarm
-                signal.alarm(0)
+                o_file.write("{}\n".format(process.stdout))
                 try:
-                    os.remove(reextension(file, "proof"))
-                    os.remove(reextension(file, "support"))
-                    os.remove(reextension(file, "ecnf"))
-                    os.remove(reextension(file, "cnf"))
-                    os.remove(reextension(file, "obg"))
+                    if os.path.exists(reextension(file, "proof")):
+                        os.remove(reextension(file, "proof"))
+                    if os.path.exists(reextension(file, "support")):
+                        os.remove(reextension(file, "support"))
+                    if os.path.exists(reextension(file, "ecnf")):
+                        os.remove(reextension(file, "ecnf"))
+                    if os.path.exists(reextension(file, "cnf")):
+                        os.remove(reextension(file, "cnf"))
+                    if os.path.exists(reextension(file, "obg")):
+                        os.remove(reextension(file, "obg"))
                 except:
                     pass
                 o_file.flush()
